Click selection moved to numpy
Click selection now uses the numpy arrays
Reordered how the selection works to optimize testing loop triangles. Now it tests for a face hover with the ray cast before testing loop triangles so that it only has to test the loop triangles of the hit face as opposed to all loop triangles of the whole mesh.